// frontend/src/model/generated/permission.dto.ts
// Code generated by tygo. DO NOT EDIT.
import {PermissionCategoryVal} from "./permission.entity"
//////////
// source: create_permission_dto.go

export interface CreatePermissionDTO {
  category: any /* entity.PermissionCategoryVal */;
  read: boolean;
  write: boolean;
  roleId: string /* uuid */;
}

//////////
// source: get_by_role_id_dto.go

export interface GetByRoleIdDTO {
  roleId: string /* uuid */;
}

//////////
// source: patch_permission_dto.go

export interface PatchPermissionDTO extends PermissionByIdDTO {
  write: boolean;
  read: boolean;
}

//////////
// source: permission_by_id_dto.go

export interface PermissionByIdDTO {
  ID: string /* uuid */;
}

//////////
// source: permissions_dto.go

export interface PermissionsDTO {
  id: string /* uuid */;
  category: PermissionCategoryVal;
  write: boolean;
  read: boolean;
}

//////////
// source: screate_permission_dto.go

export interface SCreatePermissionDTO {
  Category: any /* entity.PermissionCategoryVal */;
  Read: boolean;
  Write: boolean;
  RoleId: string /* uuid */;
}